Adds javascript things, through javascript.
The the user doesn't have javascript activated, they will through this commit
be able to see the listing of directories at the end of the page.

Rather than assuming that javascript is turned on, this will only add the
javascript functionality, if javascript is on. If it's not turned on, it will
fail gracefully, by not adding content that doesn't work.

This commit adds three layers of implementation: Internet Explorer methods,
W3C methods (pretty much everything except Interet Exporer) and inlining for
managing the event model. This should capture most javascript implementations,
unless the browser is really old.