Rewrote install script.
Made the install script less complex, and modified how values are stored in the
config file.

Rather than storing a subset of the local config file, the global configuration
file for the rt-magnet application just stores directories. This makes it
simpler to use it by the application and removes useless cruft that had no real
reason to be there in the firsts place.