Minimal javascript declaration.
A less verbose version of having the javascript routines seperate from the HTML
generation.

The scripts has been moved to the end of the body so they'll execute as (most)
of the body is done, and the DOM has been loaded.

Rather than using event managers, the button is given its event handler inlined
in the tag, and is has almost no browser specific needs.

The hiding of the text box is done at the end of the script, after the button
has been added. This to avoid it being hidden even if something goes wrong with
the script, or that javascript isn't even turned on.