Adds title, links and stub for license.
There's now a titel at the top, so it's clear what you're reading and a
link to github under links (so people can find updates and stuff).

Added license stub at the bottom. Since (almost) everything you make is
automatically copyrighted to you, and you alone, people aren't
technically allowed to make copies of your work until you've issued a
license for them to follow.

A simple license can be added at the bottom of the README for people to
get a quick understanding of what's required (or what rights they have)
when using the software.

It's also common practice (if not required) to provide the license in
the header of the source files.

List of (some) F/OSS licenses:
http://www.gnu.org/licenses/license-list.html
http://www.opensource.org/licenses/alphabetical/

Writings on the subject:
http://producingoss.com/en/license-quickstart.html
http://oreilly.com/openbook/osfreesoft/book/index.html

Note that some licenses aren't valid in some locations, like Public
Domain (not technically a license) not being applicable in the country
which you live.