Adds a table to fix the form layout.
Instead of using &nbsp; and <br> to align the elements (which will look different to most people due to their selection of fonts), this commit adds a table with the first column being the description (label) which links to the seconds column with the value input/selection.

Since the input of the magnet link is so damn long (consider changing to a textarea), align isn't used on the submit button, so it doesn't disappear outside the screen.