Fixes failed merge.
This goes through and fixes two of the major issues with the previous
merge commit that updated the javascript sections of the file.

The broken script tag in the header was removed. This is the old script
that was supposed to be replaced in the merge. Since the source file
changed while the bigger javascript source file was created, the merge
couldn't be automated. This means that there's a merge conflict, where
it asks the user to manually make the changes as required to achieve
whatever they wanted.

The conflicting sections are usually highlighted with a series of arrows
to highlight the conflicting section, with a series of equal signs
splitting, the available options.

Example

    What kind of fruit do you like?
    <<<<<<< HEAD
    I like pears!
    =======
    I like apples!
    >>>>>>> REMOTE
    Cool, so do I!

In the example above, there's has been a conflict in the choice. Two
different developers has answered the same question, and Git doesn't
know which one to pick. Instead of randomly picking one, it presents
the user with the option to use what's local (HEAD), or what comes from
something remote.

The idea here, is that you pick either the pear, or the apple response.
What is requested is a simple response, not a contradictory one that
tells it to do many things at once. Mixing apples with pears, is thus;
a bad idea.

A more pragmatic example:

    int d;
    <<<<<<< HEAD
    d = 1;
    =======
    d = 0;
    >>>>>>> REMOTE
    return 10 / d;

In this case, if you were to mix the two options, you would set d to 0.
This would in turn make the last action be 10/0, which would most
likely crash the application.

The above mixing is what the issue is with the script on top, and the
elements at the bottom of the page. Rather than selecting one of the
two options, both are used, which creates some adverse side effects.